New game modes added + bug fixes

+++ PomoPong/frequence_time.h
@@ -5,8 +5,8 @@ class frequence_time
 public:
 	frequence_time();
 	~frequence_time();
-	int accumulated_time;
-	int step_time;
+	float accumulated_time;
+	float step_time;
 	void update();
 	int current_time;
 private:
